feature - 사용자 보유 아이템 리스트 조회 기능 및 테스트 코드 개발

+++ src/main/java/com/example/skptemp/domain/item/entity/UserItem.java
@@ -18,10 +18,10 @@ public class UserItem extends BaseEntity {
     private Long itemId;
     private int count;
 
-    public void addItem(Long count){
+    public void addItem(int count){
         this.count += count;
     }
-    public void removeItem(Long count){
+    public void removeItem(int count){
         validate(count);
         this.count -= count;
     }
@@ -36,7 +36,7 @@ public class UserItem extends BaseEntity {
         return new UserItem(userId, itemId, 0);
     }
 
-    private void validate(Long count){
+    private void validate(int count){
         if(this.count < count) throw new GlobalException(GlobalErrorCode.ITEM_COUNT_EXCEPTION);
     }
 }
